Add attribute loading module
Norm::Attribute::Loading's purpose is to allow any class to implement
class methods for loading of various types into a new instance of
themselves. Or, at least, a new instance of something resembling
themselves. In truth, the only requirement is that the thing they return
have a #to_s method that plays nicely with Postgresql, since exec_params
always casts parameters to strings using this method.

This is why we have a load, but not a corresponding dump method. The
objects themselves do their own "dumping" via to_s.