Add Integer and Decimal attributes, and Loader.
The Loader will allow us to do stuff like this in a model, eventually:

  attribute :quantity, Attr::Integer
  attribute :price,    Attr::Decimal(10, 2)

We're explicitly not concerning ourselves with creating our own custom
objects in this case, because it's expected that we will encourage users
to use the supplied setters and getters, which will load the value they
set, casting it to the appropriate type and (in the case of Decimal)
enforcing the rounding rules. The precision rules, well, they're more of
an annotation than not, since BigDecimal doesn't really seem to enforce
them. :(