Add initialization and attributes to Record
One of the things I've always found annoying about ActiveRecord is how
ActiveRecord::Base#attributes= updates certain attributes but doesn't
actually set the attributes hash. For that matter, the fact that it uses
an attributes hash and tries to maintain "before typecast" versions of
the attributes as well as the one you really want to use in your code.
I'd much rather have the attribute be an instance variable, as you'd
most likely expect in a typical Ruby class, and avoid increasing the API
(internal and external) footprint of the Record object. With the
attribute methods in their own module we can always override them in the
class including Record, if needed, to grab the original value.