WIP: Repository result return API
Need to think through how we want to make sure that result objects are
available in the case of an exception. Is it sufficient to rescue and
return a result like this?

Ideally, the contract we would encourage for query methods on the
repository would be to always return a result object, but how do we want
to make this play nicely with transactions?