Begin to add error handling and transactions
Turns out I was stressing over nothing when it came to worrying about
the context handling. As it turns out, ConnectionPool always returns the
same object to a single thread. That will pose a minor hassle for people
who want to check out another connection to do some parallel-type
things, but there are a number of workarounds for that off the top of my
head, and frankly, if someone wants to do fairly "advanced" stuff like
that, Norm will get in the way a lot less than ActiveRecord already does
in that regard.