Naive atomicity in insert/update/delete
Going to need to refactor the mass_store and mass_update stuff to
prevent partial application of changes. The alternative is wrapping the
entire method in an atomically_on but that isn't terribly efficient
since it'll be savepoint-heavy.

Related: I'm not even entirely sure I want to continue handling
mass_update with implicit gathering of groups of identical changes. The
intention when I did this was to make all of the OOTB methods that
handle record saving/deletion have compatible method signatures. Without
this handling, mass_update becomes the odd man out. Maybe that's OK? I'm
not quite sure.