Remove premature feature
Results shouldn't need to track affected rows under normal
circumstances. It was adding unnecessary complexity to the repository
code.

With this in mind, I'm less concerned about a SuccessfulResult and
UnsuccessfulResult class, but I'm thinking it might be good to get rid
of all of these "success!" calls all over the place.

Thinking next up will be providing a means to request a Result as your
response in atomically/atomically_on, and adding a "value" attribute to
Result objects so that we can capture the return value of the block if
desired.

This would replace the "handle_constraints" option that currently
exists. I can't think of a good reason to split those two things apart,
and I like the idea of ensuring that you always get the same kind of
object as a return value with this flag.