Allow requesting a result from atomic operations
This is more consistent than handle_constraints flag. It's necessary to
allow results to inherit the success and value of a result if received
as a value to be more generally useful as return values from blocks that
might themselves have their values encapsulated in a result.

In other words, we don't want results containing results containing
results. That's just silly.