Simple connection setup and reconnect.
As of right now, I'm not particularly interested in trying to do a retry
of the statement that triggered the PG::UnableToSend error. I would
reasonably assume that the error being raised means that the statement
didn't execute, but I also don't really think it's the connection's job
to decide whether or not it makes sense to retry the statement. I'm
mostly just interested in making sure the connection object itself will
be able to serve the next request, which it does.