Actually, mass_* methods on repos were all stupid
For some reason it only occurred to me this morning that the reason
mass_* methods were so fidgety is that there really isn't a reasonable
default behavior to have, here.

Some people will want to modify the records in-place:

    @posts = posts_repo.for_user(current_user)
    # Make some changes
    posts_repo.mass_update(@posts)
    # Records in @posts changed as side-effect

Others will want to return new instances of the records:

    @posts = posts_repo.for_user(current_user)
    # Make some changes
    @posts = posts_repo.mass_update(@posts)
    # Records in @posts are the new copies from the repo

The side-effect version not only trickier to build, but it's also the
thing that encourages the idea of hooking actions into the
updated!/inserted! methods on Record objects. This "everything is a
side-effect" philosophy is ever-present in ActiveRecord, and I almost
inadvertently went down the path of encouraging it again in Norm.

For now, I *think* I'm OK with doing things via side-effects on the
single-record variants, but I could be convinced otherwise.