Remove Repository#primary_keys
This gave a bit of flexibility, in that a repository could be configured
to fetch and store records based on a different concept of "identity"
than the Record class had, but this flexibility seems unnecessary, given
how easily we can subclass a Record class to provide a different way of
looking at identity.

In the end, this would have only served to confuse people.