Add RaiseDelegate to serve as a default delegate
In the case of the RecordMutationProcessor, we won't necessarily have an
opinion on how to handle a constraint error that occurs in a *_many call.

In this case, we may opt to re-raise. Alternate ways to handle this
would be to submit a RecordCollection instead of an Array for the
`records` param to these calls, and allow it to have its own error
collection. That may be the direction we go in, but I'm not willing to
bake that in without also providing a simple opt-out to request a
reraise.