Add noop_one/many to ResultProcessor
The idea here is that we want the processor to provide a behavior that
the repository can use which says "I don't actually have any query to
run for results, but I want you to do whatever it is you do in order to
give me a return value that makes sense.

In the case of the default repository/processor combo, this is
illustrated in the behavior of update when a record has no updated
attributes. The user who called update would expect that if the record
was invalid, for instance, they'd get a false result, but otherwise, a
true result.