A quick spike on check constraint processing
If we went this route, we could do something like this on the DB:

    name character varying(255) NOT NULL
      CONSTRAINT "name:between(low:3,high:64)"
        CHECK (length(name) BETWEEN 3 AND 64)

and in a locale file:

    en:
      errors:
        messages:
          between: "must be between %{low} and %{high} characters"

For nice control over formatting of the messages. There's a 63-character
limit on identifier names in default compilations of PostgreSQL, but
in practice this shouldn't really be a problem, and we can shorten the
key/opt names if needed.