PostgreSQLRepository is THE Repository.
With memory repositories no longer a stated goal (WHAT IS HE SAYING?!?)
it's a bit silly to leave the abstract repository in place. This makes
the Repository class a good deal larger, but also more straightforward,
as all the basic parts are there for the seeing. It's still just shy of
200 LOC, according to cloc.